Auto redirect user to login screen upon auth error
